split files between folders
