+++ WebDriverPractice/Pages/CarrerSearchResultsPage.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WebDriverPractice
+namespace WebDriverPractice.Pages
 {
-	internal class JobDetailsPage
-	{
-	}
+    internal class CarrerSearchResultsPage
+    {
+    }
 }
