+++ WebDriverPractice/Pages/JobDetailsPage.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WebDriverPractice
+namespace WebDriverPractice.Pages
 {
-	internal class CarrerSearchResultsPage
-	{
-	}
+    internal class JobDetailsPage
+    {
+    }
 }
